Interfaces and classes were moved to separate files
