+++ ObjectPrinting/Common/IHaveRoot.cs
@@ -1,0 +1,7 @@
+﻿namespace ObjectPrinting.Common
+{
+    public interface IHaveRoot
+    {
+        internal PrintingConfigRoot Root { get; }
+    }
+}
