+++ ObjectPrinting/Common/IPrintingPropertyConfig.cs
@@ -1,0 +1,5 @@
+﻿namespace ObjectPrinting.Common
+{
+    public interface IPrintingPropertyConfig<TOwner, T> : IPrintingConfig<TOwner>, IPrintingPropertyBaseConfig<TOwner, T>
+    { }
+}
