+++ ObjectPrinting/ObjectPrinter.cs
@@ -1,6 +1,12 @@
+﻿using ObjectPrinting.Common;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+
 namespace ObjectPrinting
 {
-    public class ObjectPrinter
+    public static class ObjectPrinter
     {
         public static IPrintingConfig<T> For<T>()
         {
